Add files via upload
"Hospital Readmissions Reduction Program: An Economic and Operational Analysis" is the reference of this project.
The Hospital Readmissions Reduction Program (HRRP), a part of the U.S. Patient Protection and Affordable Care Act, requires the Centers for Medicare and Medicaid Services to penalize hospitals with excess readmissions. 
For hospitals, high readmission rate will give a high revenue but low patients' satisfactory level. If a penalty fee towards readmission rate is adopted, the hospitals should consider the revenue earned by readmission and the penalty fee caused by high readmission.
Now HRRP is gonna raise a 3% penalty fee to reduce the readmission rate, and they wanna know how the policy affect each hospital's action, that is, to know the effectiveness of this policy.
Here are some assumptions:
1. hospitals have no intention to rise their readmission rate to make money
2. each patient is identical
3. HRRP only affects Medicare payments and not private payments